moved path to datatypes. fixed a bug in all_paths
